I have added the code written in all the comments and fixed it
