+++ packages/mq_home_repository/lib/src/src.dart
@@ -4,8 +4,13 @@ export 'entity/mq_home_banner_entity.dart';
 export 'entity/mq_story_entity.dart';
 // models
 export 'model/mq_donation_page_response.dart';
+export 'model/mq_hatim_creator_model.dart';
+export 'model/mq_hatim_model.dart';
+export 'model/mq_hatim_participants_detail_model.dart';
 export 'model/mq_home_banner_response.dart';
+export 'model/mq_home_invited_hatims_model.dart';
 export 'model/mq_home_model.dart';
+export 'model/mq_home_user_model.dart';
 export 'model/mq_story_model_response.dart';
 // repositories
 export 'repository/mq_home_repository.dart';
